+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - UCR Unblocker now uses the current directory as default
+- Sensitivity to Axis Merger plugin added
 
 ### Fixed